Remove CompetitionIndex
